.agent/templates/api-server/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.26.0

# Linting
ruff>=0.2.0
